requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
azure-storage-blob
python-quickbooks
requests
intuit-oauth